move context wrapper to seperate file
