+++ pluginmanager/plugin_wrapper_service_v2.go
@@ -14,9 +14,7 @@
 package pluginmanager
 
 import (
-	"github.com/alibaba/ilogtail/pkg/models"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
-	"github.com/alibaba/ilogtail/pkg/selfmonitor"
 )
 
 type ServiceWrapperV2 struct {
@@ -35,98 +33,4 @@ func (wrapper *ServiceWrapperV2) Init(pluginMeta *pipeline.PluginMeta) error {
 func (wrapper *ServiceWrapperV2) StartService(pipelineContext pipeline.PipelineContext) error {
 	wrapper.pipelineCtxWrapper = newPipelineContextWrapper(pipelineContext, wrapper.outEventsTotal, wrapper.outEventGroupsTotal, wrapper.outSizeBytes)
 	return wrapper.Input.StartService(wrapper.pipelineCtxWrapper)
-}
-
-// PipelineContextWrapper is a wrapper of pipeline.PipelineContext
-// When call Collector(), it replaces the original pipeline.PipelineCollector with PipelineCollectorWrapper
-// for counting the traffic of the plugin
-type PipelineContextWrapper struct {
-	PipelineCollectorWrapper
-}
-
-func (wrapper *PipelineContextWrapper) Collector() pipeline.PipelineCollector {
-	return wrapper.PipelineCollectorWrapper
-}
-
-// PipelineCollectorWrapper is a wrapper of pipeline.PipelineCollector with extra telemetry metrics.
-type PipelineCollectorWrapper struct {
-	inner pipeline.PipelineCollector
-
-	outEventsTotal      selfmonitor.CounterMetric
-	outEventGroupsTotal selfmonitor.CounterMetric
-	outSizeBytes        selfmonitor.CounterMetric
-}
-
-// Collect single group and events data belonging to this group
-func (collectorWrapper PipelineCollectorWrapper) Collect(groupInfo *models.GroupInfo, eventList ...models.PipelineEvent) {
-	if collectorWrapper.inner == nil {
-		return
-	}
-
-	collectorWrapper.outEventsTotal.Add(int64(len(eventList)))
-	collectorWrapper.outEventGroupsTotal.Add(1)
-	size := int64(0)
-	for _, event := range eventList {
-		size += event.GetSize()
-	}
-	collectorWrapper.outSizeBytes.Add(size)
-	collectorWrapper.inner.Collect(groupInfo, eventList...)
-}
-
-// CollectList collect GroupEvents list that have been grouped
-func (collectorWrapper PipelineCollectorWrapper) CollectList(groupEventsList ...*models.PipelineGroupEvents) {
-	if collectorWrapper.inner == nil {
-		return
-	}
-	var eventCount int64
-	var size int64
-	for _, groupEvents := range groupEventsList {
-		eventCount += groupEvents.GetEventCount()
-		size += groupEvents.GetSize()
-	}
-	collectorWrapper.outEventGroupsTotal.Add(int64(len(groupEventsList)))
-	collectorWrapper.outEventsTotal.Add(eventCount)
-	collectorWrapper.outSizeBytes.Add(size)
-	collectorWrapper.inner.CollectList(groupEventsList...)
-}
-
-// ToArray returns an array containing all of the PipelineGroupEvents in this collector.
-func (collectorWrapper PipelineCollectorWrapper) ToArray() []*models.PipelineGroupEvents {
-	if collectorWrapper.inner == nil {
-		return nil
-	}
-	return collectorWrapper.inner.ToArray()
-}
-
-// Observe returns a chan that can consume PipelineGroupEvents from this collector.
-func (collectorWrapper PipelineCollectorWrapper) Observe() chan *models.PipelineGroupEvents {
-	if collectorWrapper.inner == nil {
-		return nil
-	}
-	return collectorWrapper.inner.Observe()
-}
-
-// Close closes the collector.
-func (collectorWrapper PipelineCollectorWrapper) Close() {
-	if collectorWrapper.inner == nil {
-		return
-	}
-	collectorWrapper.inner.Close()
-}
-
-func newPipelineContextWrapper(pipelineContext pipeline.PipelineContext, outEventsTotal selfmonitor.CounterMetric, outEventGroupsTotal selfmonitor.CounterMetric, outSizeBytes selfmonitor.CounterMetric) *PipelineContextWrapper {
-	w := &PipelineContextWrapper{}
-
-	var innerCollector pipeline.PipelineCollector
-	if pipelineContext != nil {
-		innerCollector = pipelineContext.Collector()
-	}
-
-	w.PipelineCollectorWrapper = PipelineCollectorWrapper{
-		inner:               innerCollector,
-		outEventsTotal:      outEventsTotal,
-		outEventGroupsTotal: outEventGroupsTotal,
-		outSizeBytes:        outSizeBytes,
-	}
-	return w
 }
